requirements.txt
numpy
pandas
requests
openai
python-dotenv
yfinance
matplotlib
colorama
gtts